Change script mode, valid link, sensible defaults
Make the two scripts executable. Git remembers the modes (permissions) of every file. Check if the link provided starts with "https://www.youtube.com" which acts as a basic validity test. Provide sensible defaults for the case where we want to extract a portion of the video/audio. So, that the script does not malfunction in case no input value is provided for variables : start_time, length and name_output