Podman/Docker testing
Certain distrubutions have deprecated versions of yt-dlp. Running the application in a podman/docker container is a sensible way of solving this dependency problem.

Alpine linux image has been chosen as its extremely lightweight.

Since, the command for running the application is very long, it can be
aliased to a shorter name